relocate root node
