+++ tests/test_ports.cpp
@@ -1,4 +1,4 @@
-#include "ast/design_file.hpp"
+#include "ast/nodes/design_file.hpp"
 #include "builder/assembly/node_builder.hpp"
 #include "builder/translator.hpp"
 #include "builder/visitor.hpp"
